rust: properly separate cli from library in vault init
