+++ rust/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "nitor-vault"
-version = "1.0.0"
+version = "1.1.0"
 edition = "2021"
 description = "Encrypted AWS key-value storage utility."
 license = "Apache-2.0"
@@ -14,13 +14,13 @@ authors = [
 [dependencies]
 aes-gcm = "0.10.3"
 anyhow = "1.0.89"
-aws-config = { version = "1.4.0", features = ["behavior-version-latest"] }
-aws-sdk-cloudformation = "1.30.0"
-aws-sdk-kms = "1.26.0"
-aws-sdk-s3 = "1.29.0"
+aws-config = { version = "1.5.8", features = ["behavior-version-latest"] }
+aws-sdk-cloudformation = "1.50.0"
+aws-sdk-kms = "1.46.0"
+aws-sdk-s3 = "1.54.0"
 aws-sdk-sts = { version = "1.45.0", features = ["behavior-version-latest"] }
 base64 = "0.22.1"
-clap = { version = "4.5.19", features = ["derive", "env"] }
+clap = { version = "4.5.20", features = ["derive", "env"] }
 colored = "2.1.0"
 rand = "0.8.5"
 serde = { version = "1.0.210", features = ["derive"] }
